Move dashboard to module
